+++ docs/changelog.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add border and tools parameter to view window
 - Add language parameter to view window
 - Build docs in SCn format
+- Add optional buttons to delete function
 
 ### Fixed
 - Main and system identifier translation with current language in scg-editor